Added UI & colision management
The "Menu" is now in a new scene (which show the Greece scene when we
click play for the second time).

Collision:
The collect of a star now increment a counter which can be seen on the
screen (bottom left).
The collision with an obstacle decrease (min 0) the counter and launch
an animation.

Optimisation of the collider on the rocks.